pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "RAG application built with LangChain"
requires-python = ">=3.12"
dependencies = [
    "langchain==1.0.7",
    "langchain-chroma==1.0.0",
    "langchain-community==0.4.1",
    "langchain-huggingface==1.0.1",
    "langchain-ollama==1.0.0",
    "model2vec==0.7.0",
    "platformdirs==4.2.2",
    "pyyaml==6.0.3",
    "sentence-transformers==5.1.2",
    "streamlit==1.51.0",
    "ipykernel==7.1.0",
    "pytrec-eval==0.5",
    "rank_bm25==0.2.2"
]

[project.optional-dependencies]
dev = [
    "ipykernel>=7.1.0",
]